可以实现删除和修改书本

+++ src/main/java/com/huihui/LMS/serviceImpl/BookServiceImpl.java
@@ -12,10 +12,12 @@ import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
+import org.springframework.util.ObjectUtils;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.huihui.LMS.dao.BookDao;
 import com.huihui.LMS.pojo.Book;
+import com.huihui.LMS.pojo.Category;
 import com.huihui.LMS.service.BookService;
 
 @Service
@@ -49,11 +51,35 @@ public class BookServiceImpl implements BookService{
 		@SuppressWarnings("deprecation")
 		Book bk = bookDao.getById(id);
 		try {
-			Files.delete(root.resolve(bk.getCover()));
+			if(!(bk.getCover()==null)) {
+				Files.delete(root.resolve(bk.getCover()));
+			}
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
 		bookDao.deleteById(id);
+		
+	}
+
+	@Override
+	public Book findById(Integer id) {
+		try {
+			Optional<Book> optional = bookDao.findById(id);
+			if(optional.isPresent()) {
+				return optional.get();
+			}
+		}catch(Exception ex) {
+			ex.printStackTrace();
+		}
+		return null;
+	}
+
+	@Override
+	public void updateBook(Book book) {
+		System.out.println("inserviceimpl"+book.getTitle());
+		System.out.println("thisistheid"+book.getId());
+		bookDao.updateBook(book.getId(),book.getIsbn(),book.getTitle(),book.getPress(),
+				book.getCover(),book.getAuthor(),book.getAmount(),book.getCategory());
 		
 	}
